+++ firebase-auth-flutterfire-ui/complete/lib/app.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+import 'auth_gate.dart';
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      theme: ThemeData(
+        primarySwatch: Colors.blue,
+      ),
+      home: const AuthGate(),
+    );
+  }
+}
