+++ firebase-auth-flutterfire-ui/complete/lib/home.dart
@@ -1,0 +1,40 @@
+import 'package:flutter/material.dart';
+import 'package:flutterfire_ui/auth.dart';
+
+class HomeScreen extends StatelessWidget {
+  const HomeScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        actions: [
+          IconButton(
+            icon: const Icon(Icons.person),
+            onPressed: () {
+              Navigator.push(
+                context,
+                MaterialPageRoute<ProfileScreen>(
+                  builder: (context) => const ProfileScreen(),
+                ),
+              );
+            },
+          )
+        ],
+        automaticallyImplyLeading: false,
+      ),
+      body: Center(
+        child: Column(
+          children: [
+            Image.asset('dash.png'),
+            Text(
+              'Welcome!',
+              style: Theme.of(context).textTheme.displaySmall,
+            ),
+            const SignOutButton(),
+          ],
+        ),
+      ),
+    );
+  }
+}
